Changed glm template argument to use double precision floating point numbers

+++ Game/source/math/matrix4x4.h
@@ -40,10 +40,10 @@ public:
 	static matrix4x4 view(const class vector3d& position, const class rotator& inRotation);
 
 	// Computes a perspective matrix from input parameters
-	static matrix4x4 perspective(float fieldOfViewDegrees, float viewWidth, float viewHeight, float nearClipPlane, float farClipPlane);
+	static matrix4x4 perspective(double fieldOfViewDegrees, double viewWidth, double viewHeight, double nearClipPlane, double farClipPlane);
 
 	// Computes an orthographic matrix from input parameters. Scale width and height to zoom the view in/out
-	static matrix4x4 orthographic(float width, float height, float nearClipPlane, float farClipPlane);
+	static matrix4x4 orthographic(double width, double height, double nearClipPlane, double farClipPlane);
 
 public:
 	void inverseInPlace();
